requirements.txt
Pillow==10.4.0
rembg==2.0.59
streamlit==1.38.0
